server_side/HotelServices/requirement.txt
alembic==0.8.2
aniso8601==1.0.0
blueprint==3.4.2
Flask==0.10.1
Flask-HTTPAuth==2.7.0
Flask-Limiter==0.8.5
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.8
limits==1.0.9
Mako==1.0.2
MarkupSafe==0.23
psycopg2==2.6.1
python-editor==0.4
pytz==2015.6
PyYAML==3.11
six==1.10.0
SQLAlchemy==1.0.8
Werkzeug==0.10.4
wheel==0.24.0
